requirements.txt
Django==1.11.7
django-filter==1.1.0
django-nose==1.4.5
djangorestframework==3.7.3
httpie==0.9.9
pytz==2017.3
